backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart